+++ packages/main/src/Avatar.ts
@@ -70,7 +70,8 @@ import "@ui5/webcomponents-icons/dist/employee.js";
 })
 /**
 * Fired on mouseup, space and enter if avatar is interactive
-*
+* <b>Note:</b> The event will not be fired if the <code>disabled</code>
+* property is set to <code>true</code>.
 * @event
 * @private
 * @since 1.0.0-rc.11
@@ -78,7 +79,22 @@ import "@ui5/webcomponents-icons/dist/employee.js";
 @event("click")
 class Avatar extends UI5Element implements ITabbable {
 	/**
+	 * Defines whether the component is disabled.
+	 * A disabled component can't be pressed or
+	 * focused, and it is not in the tab chain.
+	 *
+	 * @type {boolean}
+	 * @name sap.ui.webc.main.Avatar.prototype.disabled
+	 * @defaultvalue false
+	 * @public
+	 */
+	@property({ type: Boolean })
+	disabled!: boolean;
+
+	/**
 	 * Defines if the avatar is interactive (focusable and pressable).
+	 * <b>Note:</b> This property won't have effect if the <code>disabled</code>
+	 * property is set to <code>true</code>.
 	 * @type {boolean}
 	 * @name sap.ui.webc.main.Avatar.prototype.interactive
 	 * @defaultValue false
@@ -93,6 +109,13 @@ class Avatar extends UI5Element implements ITabbable {
 	 */
 	@property({ type: Boolean })
 	focused!: boolean;
+
+	/**
+	 * Indicates if the elements is pressed
+	 * @private
+	 */
+	@property({ type: Boolean })
+	pressed!: boolean;
 
 	/**
 	 * Defines the name of the UI5 Icon, that will be displayed.
@@ -292,7 +315,7 @@ class Avatar extends UI5Element implements ITabbable {
 	}
 
 	get tabindex() {
-		return this._tabIndex || (this.interactive ? "0" : "-1");
+		return this._tabIndex || (this._interactive ? "0" : "-1");
 	}
 
 	/**
@@ -320,11 +343,15 @@ class Avatar extends UI5Element implements ITabbable {
 	}
 
 	get _role() {
-		return this.interactive ? "button" : undefined;
+		return this._interactive ? "button" : undefined;
 	}
 
 	get _ariaHasPopup() {
 		return this._getAriaHasPopup();
+	}
+
+	get _interactive() {
+		return this.interactive && !this.disabled;
 	}
 
 	get validInitials() {
@@ -357,7 +384,7 @@ class Avatar extends UI5Element implements ITabbable {
 	 }
 
 	onBeforeRendering() {
-		this._onclick = this.interactive ? this._onClickHandler.bind(this) : undefined;
+		this._onclick = this._interactive ? this._onClickHandler.bind(this) : undefined;
 	}
 
 	async onAfterRendering() {
@@ -403,16 +430,16 @@ class Avatar extends UI5Element implements ITabbable {
 	_onClickHandler(e: MouseEvent) {
 		// prevent the native event and fire custom event to ensure the noConfict "ui5-click" is fired
 		e.stopPropagation();
-		this.fireEvent("click");
+		this._fireClick();
 	}
 
 	_onkeydown(e: KeyboardEvent) {
-		if (!this.interactive) {
+		if (!this._interactive) {
 			return;
 		}
 
 		if (isEnter(e)) {
-			this.fireEvent("click");
+			this._fireClick();
 		}
 
 		if (isSpace(e)) {
@@ -421,9 +448,14 @@ class Avatar extends UI5Element implements ITabbable {
 	}
 
 	_onkeyup(e: KeyboardEvent) {
-		if (this.interactive && !e.shiftKey && isSpace(e)) {
-			this.fireEvent("click");
+		if (this._interactive && !e.shiftKey && isSpace(e)) {
+			this._fireClick();
 		}
+	}
+
+	_fireClick() {
+		this.fireEvent("click");
+		this.pressed = !this.pressed;
 	}
 
 	_onfocusout() {
@@ -431,13 +463,13 @@ class Avatar extends UI5Element implements ITabbable {
 	}
 
 	_onfocusin() {
-		if (this.interactive) {
+		if (this._interactive) {
 			this.focused = true;
 		}
 	}
 
 	_getAriaHasPopup() {
-		if (!this.interactive || this.ariaHaspopup === "") {
+		if (!this._interactive || this.ariaHaspopup === "") {
 			return;
 		}